requirements.txt
Flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.3
python-dotenv==1.0.0
pillow==9.4.0 